alfa-test/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
flasgger==0.9.7.1
python-dateutil==2.8.2
PyJWT==2.8.0
bcrypt==4.1.2